moved subject metadata JSON creation to io. reformated MetaFile -> SubjectMetadata
